Move duplicate ISO code check into CountryValidator
Per review, consolidate the country validation logic: the ISO code
uniqueness check now lives in CountryValidator::validate() alongside the
field validation, instead of being a separate guard in the repository.
